backend/requirements.txt
fastapi==0.123.5
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
python-dotenv==1.2.1
email-validator==2.3.0


